Address review feedback on unread surfaces, mute write ordering, and locale copy

+++ app/src/main/java/dev/ipf/whitenoise/android/core/MessageProjector.kt
@@ -85,7 +85,7 @@ data class MessageTextCopy(
                 ChatListAttachmentKindFfi.FILE -> mediaDocument
                 ChatListAttachmentKindFfi.MIXED -> mediaAlbum
             }
-        return if (count > 1u) String.format(mediaCountedFormat, label, count.toInt()) else label
+        return if (count > 1u) String.format(mediaCountedFormat, label, count.toLong()) else label
     }
 
     companion object {
